pyproject.toml
[project]
name = "hive-json-poster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "kguelseven", email = "kguelseven@hey.com" }
]
requires-python = ">=3.13"
dependencies = [
    "beem>=0.24.26",
    "python-dotenv>=1.1.1",
]

[project.scripts]
hive-json-poster = "hive_json_poster:main"

[build-system]
requires = ["uv_build>=0.9.4,<0.10.0"]
build-backend = "uv_build"
